move most of the single threaded benchmark into common.cpp
